requirements.txt
# Start of Selection
# Start Generation Here
streamlit
google-generativeai
PyPDF2
pandas
scikit-learn
pyzotero
# End Generation Here
# Additional dependencies for future enhancements
# End of Selection
